+++ src/clues.ts
@@ -9,6 +9,7 @@ import imageLink from "./clues/image-link/index.js";
 import noAlt from "./clues/no-alt/index.js";
 import notOnlySpace from "./clues/not-only-space/index.js";
 import altStartWith from "./clues/start-with/index.js";
+import { Alt } from "./index.js";
 
 export interface Rule {
   /** The name of the rule */
@@ -53,3 +54,32 @@ export const clues: Clue = {
   decorative: decorative(),
   avoidEmoji: avoidEmoji(),
 };
+
+export function checkClue(alt: Alt) {
+  return Object.keys(clues).reduce((arr: string[], item: string) => {
+    const clue: Rule = clues[item];
+    arr = [...arr, ...filterClues({ ruleName: item, clue: clue, alt })];
+    return arr;
+  }, []);
+}
+
+export function filterClues({
+  ruleName,
+  clue,
+  alt,
+}: {
+  ruleName: string;
+  clue: Rule;
+  alt: Alt;
+}) {
+  if (!clue.rules) return [];
+  return clue.rules.reduce((arr: string[], item: string) => {
+    if (clue.fn && clue.fn(item, alt))
+      arr.push(createSuggestion(ruleName, item));
+    return arr;
+  }, []);
+}
+
+export function createSuggestion(ruleName: string, value?: string | number) {
+  return `${clues[ruleName].suggestion(value)} (${clues[ruleName].docs}).`;
+}
