+++ src/clues/end-punctuation/index.ts
@@ -1,5 +1,5 @@
-import { Rule } from "../../clues.js";
-import { Alt, createSuggestion } from "../../index.js";
+import { Rule, createSuggestion } from "../../clues.js";
+import { Alt } from "../../index.js";
 
 export function checkPunctuation(alt: Alt) {
   return !/[.!?]$/.test(alt) && alt.length > 1
