+++ src/clues/exact-match/__tests__/index.test.ts
@@ -1,4 +1,4 @@
-import { checkClue } from "../../../index.js";
+import { checkClue } from "../../../clues.js";
 
 test("exactMatch", () => {
   expect(checkClue("logo")).toMatchInlineSnapshot(`
