+++ src/index.ts
@@ -1,4 +1,4 @@
-import { clues, Rule } from "./clues.js";
+import { checkClue, createSuggestion } from "./clues.js";
 import { checkEmoji } from "./clues/avoid-emoji/index.js";
 import { checkLength } from "./clues/char-length/index.js";
 import { checkPunctuation } from "./clues/end-punctuation/index.js";
@@ -23,33 +23,4 @@ export default function altText(alt?: Alt) {
     ...checkEmoji(alt),
   ];
   return suggestion.length ? suggestion.join(" ") : undefined;
-}
-
-export function checkClue(alt: Alt) {
-  return Object.keys(clues).reduce((arr: string[], item: string) => {
-    const clue: Rule = clues[item];
-    arr = [...arr, ...filterClues({ ruleName: item, clue: clue, alt })];
-    return arr;
-  }, []);
-}
-
-export function filterClues({
-  ruleName,
-  clue,
-  alt,
-}: {
-  ruleName: string;
-  clue: Rule;
-  alt: Alt;
-}) {
-  if (!clue.rules) return [];
-  return clue.rules.reduce((arr: string[], item: string) => {
-    if (clue.fn && clue.fn(item, alt))
-      arr.push(createSuggestion(ruleName, item));
-    return arr;
-  }, []);
-}
-
-export function createSuggestion(ruleName: string, value?: string | number) {
-  return `${clues[ruleName].suggestion(value)} (${clues[ruleName].docs}).`;
 }
